+++ README.md
@@ -1,1 +1,1 @@
-# deup
+# DEUP
